requirements.txt
boto3
jupyter
pyspark==3.5.1
python-dotenv
pandas
pyarrow
requests
pynessie